+++ web/usuaris/src/__tests__/hooks/useGetUsers.test.tsx
@@ -30,7 +30,7 @@ describe('useGetUsers', () => {
         name: 'Marc Bofill',
         dni: '87654321B',
         status: 'ACTIVE',
-        createdAt: '2023/11/05 00:00:00.000',
+        createdAt: '2023/11/06 00:00:00.000',
         itineraryName: 'Frontend React',
       },
       {
@@ -38,7 +38,7 @@ describe('useGetUsers', () => {
         name: 'Montserrat Capdevila',
         dni: '45678912C',
         status: 'BLOCKED',
-        createdAt: '2023/11/05 00:00:00.000',
+        createdAt: '2023/11/07 00:00:00.000',
         itineraryName: 'Fullstack Php',
       },
       {
@@ -46,7 +46,7 @@ describe('useGetUsers', () => {
         name: 'Anna Brull',
         dni: '45678912D',
         status: 'BLOCKED',
-        createdAt: '2023/11/05 00:00:00.000',
+        createdAt: '2023/11/08 00:00:00.000',
         itineraryName: 'Frontend React',
       },
       {
@@ -54,7 +54,7 @@ describe('useGetUsers', () => {
         name: 'Marc Serra',
         dni: '12378912D',
         status: 'BLOCKED',
-        createdAt: '2023/11/05 00:00:00.000',
+        createdAt: '2023/11/09 00:00:00.000',
         itineraryName: 'Frontend Angular',
       },
     ])