+++ web/usuaris/src/__tests__/organisms/FiltersWidget.test.tsx
@@ -1,11 +1,11 @@
 import { fireEvent, render, screen, waitFor } from '../test-utils'
 import { FiltersWidget } from '../../components/organisms'
 
-const mockHandleFilters = vi.fn()
+const mockSetFilters = vi.fn()
 
 describe('FiltersWidget', () => {
   it('renders correctly', async () => {
-    render(<FiltersWidget filters={{}} handleFilters={mockHandleFilters} />)
+    render(<FiltersWidget filters={{}} setFilters={mockSetFilters} />)
 
     await waitFor(() =>
       expect(screen.getByText(/especialitat/i)).toBeInTheDocument()
@@ -14,7 +14,7 @@ describe('FiltersWidget', () => {
   })
 
   it('shows itinerary selected in dropdown and sends it to parent', async () => {
-    render(<FiltersWidget filters={{}} handleFilters={mockHandleFilters} />)
+    render(<FiltersWidget filters={{}} setFilters={mockSetFilters} />)
 
     await waitFor(() =>
       expect(screen.getByText(/especialitat/i)).toBeInTheDocument()
@@ -34,7 +34,7 @@ describe('FiltersWidget', () => {
     expect(dropdownHeader).toHaveTextContent('Frontend React')
     expect(screen.queryByText(/especialitat/i)).not.toBeInTheDocument()
 
-    expect(mockHandleFilters).toHaveBeenCalledWith({
+    expect(mockSetFilters).toHaveBeenCalledWith({
       itinerarySlug: 'react',
     })
   })
@@ -43,7 +43,7 @@ describe('FiltersWidget', () => {
     render(
       <FiltersWidget
         filters={{ status: 'ACTIVE' }}
-        handleFilters={mockHandleFilters}
+        setFilters={mockSetFilters}
       />
     )
 
@@ -62,24 +62,24 @@ describe('FiltersWidget', () => {
 
     expect(dropdownHeader).toHaveTextContent('Frontend React')
 
-    expect(mockHandleFilters).toHaveBeenCalledWith({
+    expect(mockSetFilters).toHaveBeenCalledWith({
       itinerarySlug: 'react',
       status: 'ACTIVE',
     })
 
     fireEvent.click(dropdownHeader)
 
-    const deselectReact = screen.getByTestId('deselect-1')
+    const deselectReact = screen.getAllByTestId('deselect-1')
 
-    fireEvent.click(deselectReact)
+    fireEvent.click(deselectReact[0])
 
-    expect(mockHandleFilters).toHaveBeenCalledWith({
+    expect(mockSetFilters).toHaveBeenCalledWith({
       status: 'ACTIVE',
     })
   })
 
   it('displays search bar and updates filters value when typing', () => {
-    render(<FiltersWidget filters={{}} handleFilters={mockHandleFilters} />)
+    render(<FiltersWidget filters={{}} setFilters={mockSetFilters} />)
 
     const searchBar = screen.getByPlaceholderText(/Cercar/i)
     expect(searchBar).toBeInTheDocument()
@@ -87,7 +87,7 @@ describe('FiltersWidget', () => {
     fireEvent.change(searchBar, { target: { value: 'marc' } })
 
     waitFor(() =>
-      expect(mockHandleFilters).toHaveBeenCalledWith({
+      expect(mockSetFilters).toHaveBeenCalledWith({
         name: 'marc',
         dni: 'marc',
       })
@@ -98,7 +98,7 @@ describe('FiltersWidget', () => {
     render(
       <FiltersWidget
         filters={{ status: 'ACTIVE' }}
-        handleFilters={mockHandleFilters}
+        setFilters={mockSetFilters}
       />
     )
 
@@ -106,7 +106,7 @@ describe('FiltersWidget', () => {
     fireEvent.change(searchBar, { target: { value: 'marc' } })
 
     waitFor(() =>
-      expect(mockHandleFilters).toHaveBeenCalledWith({
+      expect(mockSetFilters).toHaveBeenCalledWith({
         status: 'ACTIVE',
         name: 'marc',
         dni: 'marc',
@@ -117,9 +117,64 @@ describe('FiltersWidget', () => {
     fireEvent.click(closeButton)
 
     waitFor(() =>
-      expect(mockHandleFilters).toHaveBeenCalledWith({
+      expect(mockSetFilters).toHaveBeenCalledWith({
         status: 'ACTIVE',
       })
     )
+  })
+
+  it('updates filters value when introducing dates', () => {
+    render(<FiltersWidget filters={{}} setFilters={mockSetFilters} />)
+
+    const startDateInput = screen.getByPlaceholderText('De...')
+    const endDateInput = screen.getByPlaceholderText('Fins...')
+
+    const mockStartDate = new Date()
+    const mockEndDate = new Date()
+
+    fireEvent.change(startDateInput, { target: { value: mockStartDate } })
+    fireEvent.change(endDateInput, { target: { value: mockEndDate } })
+
+    waitFor(() =>
+      expect(mockSetFilters).toHaveBeenCalledWith({
+        startDate: mockStartDate.toISOString(),
+        endDate: mockEndDate.toISOString(),
+      })
+    )
+  })
+
+  it('deletes dates when closing dates inputs and updates filters accordingly', () => {
+    render(
+      <FiltersWidget
+        filters={{ status: 'ACTIVE' }}
+        setFilters={mockSetFilters}
+      />
+    )
+
+    const startDateInput = screen.getByPlaceholderText('De...')
+    const endDateInput = screen.getByPlaceholderText('Fins...')
+
+    const mockStartDate = new Date()
+    const mockEndDate = new Date()
+
+    fireEvent.change(startDateInput, { target: { value: mockStartDate } })
+    fireEvent.change(endDateInput, { target: { value: mockEndDate } })
+
+    waitFor(() =>
+      expect(mockSetFilters).toHaveBeenCalledWith({
+        status: 'ACTIVE',
+        startDate: mockStartDate.toISOString(),
+        endDate: mockEndDate.toISOString(),
+      })
+    )
+
+    const closeIcon = screen.getByTestId('close-button-test')
+    fireEvent.click(closeIcon)
+
+    waitFor(() => {
+      expect(mockSetFilters).toHaveBeenCalledWith({
+        status: 'ACTIVE',
+      })
+    })
   })
 })