+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.31.7] - 2024-05-02
+
+### Fixed
+
+- Fix Dropdown to allow custom data-testid in header
+
+## [0.31.6] - 2024-05-02
+
+### Changed
+
+- Deleted unnecessary dependency in Search component
+
 ## [0.31.5] - 2024-04-24
 
 ### Changed